Moved cache types to types.go.
